make a C version of SA1inittable and remove the asm version
No compelling reason for the asm version.  All it's doing is filling in
some arrays with opcodes.  SA1inittable can be written in C just like
inittable and inittablec.

This change gets one step closer to being able to run this as a native
x86-64 binary.  The assembly SA1inittable function does some stuff with
the ES register which generates Illegal Instruction when compiled as a
64-bit binary.

Tested with two games using the SA-1 (Super Accelerator) Chip.
  * Super Mario RPG: Legend of the seven starts.
  * Kirby Super Star

Super Mario RPB plays for a little while, but seem to glitch out and
black screen with and without this change.

Kirby Super Star seems fine...