fix ao thread so it can stop on app exit
Previously the ao thread was getting killed by a SIGTERM, but it was also
deadlocked.  This change uses a terminated global, signals the condition
variable to wake up the thread so it can notice it's terminated, then the
main thread joins the audio thread and life goes on without a crash.

This change also removes the timedwait because there is no need to wake
up if there is no signal.